+++ crates/zng-var/build.rs
@@ -1,4 +1,8 @@
 fn main() {
+    if cfg!(feature = "dyn_closure") {
+        println!(r#"cargo:warning=feature "dyn_closure" is deprecated, no longer needed"#);
+    }
+
     macro_rules! enable {
         ($feature:tt) => {
             if !cfg!(feature = $feature) {